+++ oqs/src/sig.rs
@@ -5,6 +5,7 @@
 use alloc::vec::Vec;
 
 use core::ptr::NonNull;
+use core::str::FromStr;
 
 #[cfg(not(feature = "std"))]
 use cstr_core::CStr;
@@ -26,7 +27,7 @@ newtype_buffer!(Signature, SignatureRef);
 pub type Message = [u8];
 
 macro_rules! implement_sigs {
-    { $(($feat: literal) $sig: ident: $oqs_id: ident),* $(,)? } => (
+    { $(($feat: literal) $sig: ident: $oqs_id: ident: $str_name: literal),* $(,)? } => (
         /// Supported algorithms by liboqs
         ///
         /// They may not all be enabled
@@ -48,6 +49,18 @@ macro_rules! implement_sigs {
                 )*
             };
             id as *const _ as *const libc::c_char
+        }
+
+        impl FromStr for Algorithm {
+            type Err = Error;
+            fn from_str(s: &str) -> Result<Self> {
+                match s {
+                    $(
+                        $str_name => Ok(Algorithm::$sig),
+                    )*
+                    _ => Err(Error::AlgorithmNotSupportedOrKnown),
+                }
+            }
         }
 
         $(
@@ -108,29 +121,73 @@ macro_rules! implement_sigs {
                         assert!(!version.is_empty());
                     }
                 }
+
+
+                #[test]
+                fn test_from_str() {
+                    let algo = Algorithm::$sig;
+                    let name = algo.name();
+                    let res = name.parse::<Algorithm>();
+                    if res.is_err() {
+                        assert!(false, "Failed to parse: {:?}", name);
+                    }
+                    let algo2 = res.unwrap();
+                    assert_eq!(algo, algo2);
+                }
             }
         )*
     )
 }
 
+// List of supported Sig algorithms found in liboqs/src/sig/sig.h
 implement_sigs! {
-    ("dilithium") Dilithium2: OQS_SIG_alg_dilithium_2,
-    ("dilithium") Dilithium3: OQS_SIG_alg_dilithium_3,
-    ("dilithium") Dilithium5: OQS_SIG_alg_dilithium_5,
-    ("falcon") Falcon512: OQS_SIG_alg_falcon_512,
-    ("falcon") Falcon1024: OQS_SIG_alg_falcon_1024,
-    ("sphincs") SphincsSha2128fSimple: OQS_SIG_alg_sphincs_sha2_128f_simple,
-    ("sphincs") SphincsSha2128sSimple: OQS_SIG_alg_sphincs_sha2_128s_simple,
-    ("sphincs") SphincsSha2192fSimple: OQS_SIG_alg_sphincs_sha2_192f_simple,
-    ("sphincs") SphincsSha2192sSimple: OQS_SIG_alg_sphincs_sha2_192s_simple,
-    ("sphincs") SphincsSha2256fSimple: OQS_SIG_alg_sphincs_sha2_256f_simple,
-    ("sphincs") SphincsSha2256sSimple: OQS_SIG_alg_sphincs_sha2_256s_simple,
-    ("sphincs") SphincsShake128fSimple: OQS_SIG_alg_sphincs_shake_128f_simple,
-    ("sphincs") SphincsShake128sSimple: OQS_SIG_alg_sphincs_shake_128s_simple,
-    ("sphincs") SphincsShake192fSimple: OQS_SIG_alg_sphincs_shake_192f_simple,
-    ("sphincs") SphincsShake192sSimple: OQS_SIG_alg_sphincs_shake_192s_simple,
-    ("sphincs") SphincsShake256fSimple: OQS_SIG_alg_sphincs_shake_256f_simple,
-    ("sphincs") SphincsShake256sSimple: OQS_SIG_alg_sphincs_shake_256s_simple,
+    ("dilithium") Dilithium2: OQS_SIG_alg_dilithium_2: "Dilithium2",
+    ("dilithium") Dilithium3: OQS_SIG_alg_dilithium_3: "Dilithium3",
+    ("dilithium") Dilithium5: OQS_SIG_alg_dilithium_5: "Dilithium5",
+    ("falcon") Falcon512: OQS_SIG_alg_falcon_512: "Falcon-512",
+    ("falcon") Falcon1024: OQS_SIG_alg_falcon_1024: "Falcon-1024",
+    ("falcon") FalconPadded512: OQS_SIG_alg_falcon_padded_512: "Falcon-padded-512",
+    ("falcon") FalconPadded1024: OQS_SIG_alg_falcon_padded_1024: "Falcon-padded-1024",
+    ("sphincs") SphincsSha2128fSimple: OQS_SIG_alg_sphincs_sha2_128f_simple: "SPHINCS+-SHA2-128f-simple",
+    ("sphincs") SphincsSha2128sSimple: OQS_SIG_alg_sphincs_sha2_128s_simple: "SPHINCS+-SHA2-128s-simple",
+    ("sphincs") SphincsSha2192fSimple: OQS_SIG_alg_sphincs_sha2_192f_simple: "SPHINCS+-SHA2-192f-simple",
+    ("sphincs") SphincsSha2192sSimple: OQS_SIG_alg_sphincs_sha2_192s_simple: "SPHINCS+-SHA2-192s-simple",
+    ("sphincs") SphincsSha2256fSimple: OQS_SIG_alg_sphincs_sha2_256f_simple: "SPHINCS+-SHA2-256f-simple",
+    ("sphincs") SphincsSha2256sSimple: OQS_SIG_alg_sphincs_sha2_256s_simple: "SPHINCS+-SHA2-256s-simple",
+    ("sphincs") SphincsShake128fSimple: OQS_SIG_alg_sphincs_shake_128f_simple: "SPHINCS+-SHAKE-128f-simple",
+    ("sphincs") SphincsShake128sSimple: OQS_SIG_alg_sphincs_shake_128s_simple: "SPHINCS+-SHAKE-128s-simple",
+    ("sphincs") SphincsShake192fSimple: OQS_SIG_alg_sphincs_shake_192f_simple: "SPHINCS+-SHAKE-192f-simple",
+    ("sphincs") SphincsShake192sSimple: OQS_SIG_alg_sphincs_shake_192s_simple: "SPHINCS+-SHAKE-192s-simple",
+    ("sphincs") SphincsShake256fSimple: OQS_SIG_alg_sphincs_shake_256f_simple: "SPHINCS+-SHAKE-256f-simple",
+    ("sphincs") SphincsShake256sSimple: OQS_SIG_alg_sphincs_shake_256s_simple: "SPHINCS+-SHAKE-256s-simple",
+    ("ml_dsa") MlDsa44Ipd: OQS_SIG_alg_ml_dsa_44_ipd: "ML-DSA-44-ipd",
+    ("ml_dsa") MlDsa65Ipd: OQS_SIG_alg_ml_dsa_65_ipd: "ML-DSA-65-ipd",
+    ("ml_dsa") MlDsa87Ipd: OQS_SIG_alg_ml_dsa_87_ipd: "ML-DSA-87-ipd",
+    ("ml_dsa") MlDsa44: OQS_SIG_alg_ml_dsa_44: "ML-DSA-44",
+    ("ml_dsa") MlDsa65: OQS_SIG_alg_ml_dsa_65: "ML-DSA-65",
+    ("ml_dsa") MlDsa87: OQS_SIG_alg_ml_dsa_87: "ML-DSA-87",
+    ("mayo") Mayo1: OQS_SIG_alg_mayo_1: "MAYO-1",
+    ("mayo") Mayo2: OQS_SIG_alg_mayo_2: "MAYO-2",
+    ("mayo") Mayo3: OQS_SIG_alg_mayo_3: "MAYO-3",
+    ("mayo") Mayo5: OQS_SIG_alg_mayo_5: "MAYO-5",
+    ("cross") CrossRsdp128Balanced: OQS_SIG_alg_cross_rsdp_128_balanced: "cross-rsdp-128-balanced",
+    ("cross") CrossRspd128Fast: OQS_SIG_alg_cross_rsdp_128_fast: "cross-rsdp-128-fast",
+    ("cross") CrossRspd128Small: OQS_SIG_alg_cross_rsdp_128_small: "cross-rsdp-128-small",
+    ("cross") CrossRspd192Balanced: OQS_SIG_alg_cross_rsdp_192_balanced: "cross-rsdp-192-balanced",
+    ("cross") CrossRspd192Fast: OQS_SIG_alg_cross_rsdp_192_fast: "cross-rsdp-192-fast",
+    ("cross") CrossRspd192Small: OQS_SIG_alg_cross_rsdp_192_small: "cross-rsdp-192-small",
+    ("cross") CrossRspd256Balanced: OQS_SIG_alg_cross_rsdp_256_balanced: "cross-rsdp-256-balanced",
+    ("cross") CrossRspd256Fast: OQS_SIG_alg_cross_rsdp_256_fast: "cross-rsdp-256-fast",
+    ("cross") CrossRspd256Small: OQS_SIG_alg_cross_rsdp_256_small: "cross-rsdp-256-small",
+    ("cross") CrossRspdg128Balanced: OQS_SIG_alg_cross_rsdpg_128_balanced: "cross-rsdpg-128-balanced",
+    ("cross") CrossRspdg128Fast: OQS_SIG_alg_cross_rsdpg_128_fast: "cross-rsdpg-128-fast",
+    ("cross") CrossRspdg128Small: OQS_SIG_alg_cross_rsdpg_128_small: "cross-rsdpg-128-small",
+    ("cross") CrossRspdg192Balanced: OQS_SIG_alg_cross_rsdpg_192_balanced: "cross-rsdpg-192-balanced",
+    ("cross") CrossRspdg192Fast: OQS_SIG_alg_cross_rsdpg_192_fast: "cross-rsdpg-192-fast",
+    ("cross") CrossRspdg192Small: OQS_SIG_alg_cross_rsdpg_192_small: "cross-rsdpg-192-small",
+    ("cross") CrossRspdg256Balanced: OQS_SIG_alg_cross_rsdpg_256_balanced: "cross-rsdpg-256-balanced",
+    ("cross") CrossRspdg256Fast: OQS_SIG_alg_cross_rsdpg_256_fast: "cross-rsdpg-256-fast",
+    ("cross") CrossRspdg256Small: OQS_SIG_alg_cross_rsdpg_256_small: "cross-rsdpg-256-small",
 }
 
 impl Algorithm {
@@ -175,6 +232,43 @@ pub struct Sig {
     sig: NonNull<ffi::OQS_SIG>,
 }
 
+impl Debug for Sig {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Sig")
+            .field("algorithm", &self.algorithm)
+            .field(
+                "method_name",
+                &unsafe { CStr::from_ptr(self.sig.as_ref().method_name) }
+                    .to_str()
+                    .expect("method name"),
+            )
+            .field(
+                "alg_version",
+                &unsafe { CStr::from_ptr(self.sig.as_ref().alg_version) }
+                    .to_str()
+                    .expect("alg_version"),
+            )
+            .field(
+                "claimed_nist_level",
+                &unsafe { self.sig.as_ref() }.claimed_nist_level,
+            )
+            .field("euf_cma", &unsafe { self.sig.as_ref() }.euf_cma)
+            .field(
+                "length_public_key",
+                &unsafe { self.sig.as_ref() }.length_public_key,
+            )
+            .field(
+                "length_secret_key",
+                &unsafe { self.sig.as_ref() }.length_secret_key,
+            )
+            .field(
+                "length_signature",
+                &unsafe { self.sig.as_ref() }.length_signature,
+            )
+            .finish()
+    }
+}
+
 unsafe impl Sync for Sig {}
 unsafe impl Send for Sig {}
 
@@ -184,15 +278,14 @@ impl Drop for Sig {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::fmt::Display for Algorithm {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.name().fmt(f)
+impl Display for Algorithm {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        Display::fmt(self.name(), f)
     }
 }
 
-impl core::convert::TryFrom<Algorithm> for Sig {
-    type Error = crate::Error;
+impl TryFrom<Algorithm> for Sig {
+    type Error = Error;
     fn try_from(alg: Algorithm) -> Result<Sig> {
         Sig::new(alg)
     }
@@ -284,7 +377,7 @@ impl Sig {
     /// Generate a new keypair
     pub fn keypair(&self) -> Result<(PublicKey, SecretKey)> {
         let sig = unsafe { self.sig.as_ref() };
-        let func = sig.keypair.unwrap();
+        let func = sig.keypair.expect("keypair function not available");
         let mut pk = PublicKey {
             bytes: Vec::with_capacity(sig.length_public_key),
         };
@@ -309,7 +402,7 @@ impl Sig {
     ) -> Result<Signature> {
         let sk = sk.into();
         let sig = unsafe { self.sig.as_ref() };
-        let func = sig.sign.unwrap();
+        let func = sig.sign.expect("sign function not available");
         let mut sig = Signature {
             bytes: Vec::with_capacity(sig.length_signature),
         };
@@ -346,7 +439,7 @@ impl Sig {
             return Err(Error::InvalidLength);
         }
         let sig = unsafe { self.sig.as_ref() };
-        let func = sig.verify.unwrap();
+        let func = sig.verify.expect("verify function not available");
         let status = unsafe {
             func(
                 message.as_ptr(),